feat: adjust preview banner
